migrate type converters code into a module
