+++ bakertypes/src/main/scala/com/ing/baker/types/TypeModule.scala
@@ -4,9 +4,9 @@ trait TypeModule {
 
   def isApplicable(javaType: java.lang.reflect.Type): Boolean
 
-  def readType(context: TypeContext, javaType: java.lang.reflect.Type): Type
+  def readType(context: TypeConverter, javaType: java.lang.reflect.Type): Type
 
-  def toJava(context: TypeContext, value: Value, javaType: java.lang.reflect.Type): Any
+  def toJava(context: TypeConverter, value: Value, javaType: java.lang.reflect.Type): Any
 
-  def fromJava(context: TypeContext, obj: Any): Value
+  def fromJava(context: TypeConverter, obj: Any): Value
 }
