+++ bakertypes/src/main/scala/com/ing/baker/types/modules/JavaCollections.scala
@@ -4,7 +4,7 @@ import java.lang.reflect.ParameterizedType
 import java.util
 
 import com.ing.baker.types.Converters.{isAssignableToBaseClass, toValue}
-import com.ing.baker.types.{ListType, ListValue, TypeContext, TypeModule, Value}
+import com.ing.baker.types.{ListType, ListValue, TypeConverter, TypeModule, Value}
 
 import scala.collection.JavaConverters._
 
@@ -17,13 +17,13 @@ object JavaCollections {
     override def isApplicable(javaType: java.lang.reflect.Type): Boolean =
       isAssignableToBaseClass(javaType, classOf[java.util.List[_]])
 
-    override def readType(context: TypeContext, javaType: java.lang.reflect.Type) = {
+    override def readType(context: TypeConverter, javaType: java.lang.reflect.Type) = {
       val genericType = javaType.asInstanceOf[ParameterizedType]
       val entryType = context.readType(genericType.getActualTypeArguments()(0))
       ListType(entryType)
     }
 
-    override  def toJava(context: TypeContext, value: Value, javaType: java.lang.reflect.Type) = value match {
+    override  def toJava(context: TypeConverter, value: Value, javaType: java.lang.reflect.Type) = value match {
       case ListValue(entries) if isApplicable(javaType) =>
         val entryType = javaType.asInstanceOf[ParameterizedType].getActualTypeArguments()(0)
         val list = new util.ArrayList[Any]()
@@ -34,7 +34,7 @@ object JavaCollections {
         list
     }
 
-    def fromJava(context: TypeContext, obj: Any): Value = obj match {
+    def fromJava(context: TypeConverter, obj: Any): Value = obj match {
       case list: java.util.List[_] => ListValue(list.asScala.toList.map(toValue))
     }
   }
